Refactor CtaTrait out of ElementTrait
