+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphCta.php
@@ -22,6 +22,7 @@ use Drupal\server_general\ProcessedTextBuilderTrait;
 class ParagraphCta extends EntityViewBuilderPluginAbstract {
 
   use ButtonTrait;
+  use ElementTrait\CtaTrait;
   use ElementTrait;
   use ElementWrapTrait;
   use ProcessedTextBuilderTrait;
